feat: add repository pattern, database indexes, and comprehensive tests - Add TaskRepository and ScheduleRepository for clean data layer abstraction - Add database indexes on date, type, and is_imported columns for performance - Add toMapWithId() method to TaskItem for safe update operations - Add error handling and loading states to TasksScreen - Fix deprecated APIs: value→initialValue, clamp().toInt() - Fix FLAG_MUTABLE for Android widget list view support - Add sqflite_common_ffi for proper database testing - Add 30 new tests covering repositories, models, and widget tests - Add lint rules: prefer_const_constructors, sized_box_for_whitespace

+++ lib/ui/dashboard_screen.dart
@@ -63,7 +63,7 @@ class _DashboardScreenState extends ConsumerState<DashboardScreen>
             final today = DateTime.now();
             final daysToMax = maxDate.difference(today).inDays;
             // Cap at 120 days as per user's 4-month request, but always allow at least 15 days
-            final futureBuffer = daysToMax.clamp(15, 120);
+            final futureBuffer = daysToMax.clamp(15, 120).toInt();
             _totalPages = AppConstants.centerIndex + futureBuffer + 1;
           } catch (_) {
             _totalPages = AppConstants.centerIndex + 120; // fallback
